+++ src/extensions/display/displayext_hooks.cpp
@@ -406,6 +406,25 @@ DEFINE_HOOK(0x0047A856, _DisplayClass_47A790_Patch, 0)
     }
 }
 
+/**
+ *  Patches DisplayClass::Mouse_Left_Release to skip the 'Active_Click' call when the action
+ *  is ACTION_TOGGLE_SELECT (adding to selection). In most cases, this was unnecessary and did nothing.
+ * 
+ *  However, in the case of medics, since they can click on themselves to switch to Area Guard,
+ *  this caused adding them to selection to immediately move them to Area Guard, while also flashing.
+ *
+ *  @author: JoyfulShush
+ */
+DEFINE_HOOK(0x00478BC7, _Display_Class_Mouse_Left_Release_Toggle_Select_Patch, 5)
+{
+    GET(ActionType, action, EDI);
+
+    if (action == ACTION_TOGGLE_SELECT) {
+        return 0x00478BD8;
+    }
+
+    return 0;
+}
 
 /**
  *  Main function for patching the hooks.