+++ src/extensions/house/houseext_hooks.cpp
@@ -1712,6 +1712,21 @@ DEFINE_HOOK(0x004C0F87, _HouseClass_AI_Raise_Money_Fix_Memory_Corruption, 0)
     return 0x004C0F9F;
 }
 
+/**
+ *  Patches HouseClass::Recalc_Radar_Availability to allow Free Radar to still function when the player is in low power.
+ *  This requires 'FreeRadarOnLowPower=yes' to be set under [General].
+ *  Note that Ion Storms still turns off the radar regardless of this flag.
+ *
+ *  @author: JoyfulShush
+ */
+DEFINE_HOOK(0x004C958D, _HouseClass_Recalc_Radar_Availability_Free_Radar_Low_Power_Patch, 6)
+{
+    if (Scen->IsFreeRadar && RuleExtension->IsFreeRadarOnLowPower) {
+        return 0x004C966A;
+    }
+
+    return 0;
+}
 
 /**
  *  #issue-177