+++ src/extensions/rules/rulesext.cpp
@@ -104,6 +104,9 @@ RulesClassExtension::RulesClassExtension(const RulesClass* this_ptr) :
     PausedRepairsFrame(6),
     EscortRange(-1),
     AbandonTargetEscortRange(-1),
+	IsFreeRadarOnLowPower(false),
+    IsUseBridgeHealth(false),
+	BridgeArmor(ARMOR_NULL),
     BaseUnit(),
     Diff(),
     PlayerNormal(),
@@ -268,6 +271,9 @@ void RulesClassExtension::Object_CRC(CRCEngine &crc) const
     crc(PausedRepairsFrame);
     crc(EscortRange);
     crc(AbandonTargetEscortRange);
+    crc(IsFreeRadarOnLowPower);
+    crc(IsUseBridgeHealth);
+    crc(BridgeArmor);
     crc(BaseUnit.Count());
 }
 
@@ -701,6 +707,8 @@ bool RulesClassExtension::General(CCINIClass &ini)
     PausedRepairsFrame = ini.Get_Int(GENERAL, "PausedRepairsFrame", PausedRepairsFrame);
     EscortRange = ini.Get_Lepton(GENERAL, "EscortRange", EscortRange);
     AbandonTargetEscortRange = ini.Get_Lepton(GENERAL, "AbandonTargetEscortRange", AbandonTargetEscortRange);
+    IsFreeRadarOnLowPower = ini.Get_Bool(GENERAL, "FreeRadarOnLowPower", IsFreeRadarOnLowPower);
+    IsUseBridgeHealth = ini.Get_Bool(GENERAL, "UseBridgeHealth", IsUseBridgeHealth);
 
     /**
      *  Allow replacing any signle movement zone with a copy of RA2's water MZone.
@@ -791,6 +799,7 @@ bool RulesClassExtension::CombatDamage(CCINIClass & ini)
 
     IceStrength = ini.Get_Int(COMBATDAMAGE, "IceStrength", IceStrength);
     BuildingFlameSpawnBlockFrames = ini.Get_Int(COMBATDAMAGE, "BuildingFlameSpawnBlockFrames", BuildingFlameSpawnBlockFrames);
+    BridgeArmor = ini.Get_ArmorType(COMBATDAMAGE, "BridgeArmor", BridgeArmor);
 
     return true;
 }