+++ src/vinifera/vinifera_globals.cpp
@@ -97,3 +97,5 @@ bool Vinifera_SkipToInternet = false;
 bool Vinifera_ExitAfterSkip = false;
 
 DynamicVectorClass<ExceptionInfoDatabaseStruct> ExceptionInfoDatabase;
+
+std::unordered_map<Cell, int, CellHasher> BridgeHealths;